Add Storage API support
I had some availability so I decided to make some effort in order to get
the Storage API working.

There is a BC Break because it looks like a storage object is already
implemented for the plan. So I am not sure how to proceed. But that's
where I am so far.

Related #3

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>